requirements.txt

# Additional requirements go here

# Configuration
django-environ==0.4.4

telepot

django-annoying

ujson